requirements.txt
atomicwrites==1.4.0
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
cycler==0.10.0
debugpy==1.3.0
decorator==5.0.9
idna==3.2
iniconfig==1.1.1
ipykernel==6.0.1
ipython==7.25.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.12
jupyter-core==4.7.1
kiwisolver==1.3.1
lxml==4.6.3
matplotlib==3.4.2
matplotlib-inline==0.1.2
multitasking==0.0.9
numpy==1.21.1
packaging==21.0
pandas==1.3.1
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.1
pluggy==0.13.1
prompt-toolkit==3.0.19
py==1.10.0
Pygments==2.9.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
python-decouple==3.4
pytz==2021.1
pywin32==301
pyzmq==22.1.0
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
toml==0.10.2
tornado==6.1
traitlets==5.0.5
urllib3==1.26.6
wcwidth==0.2.5
wincertstore==0.2
yfinance==0.1.63
